Action Refactor Part #2
