+++ module/data/action/effectAction.mjs
@@ -1,7 +1,7 @@
 import DHBaseAction from './baseAction.mjs';
 
 export default class DHEffectAction extends DHBaseAction {
-    static extraSchemas = ['effects', 'target'];
+    static extraSchemas = [...super.extraSchemas, 'effects', 'target'];
 
     async trigger(event, data) {
         if(this.effects.length) {
