+++ module/data/fields/_module.mjs
@@ -3,3 +3,4 @@ export { default as FormulaField } from './formulaField.mjs';
 export { default as ForeignDocumentUUIDField } from './foreignDocumentUUIDField.mjs';
 export { default as ForeignDocumentUUIDArrayField } from './foreignDocumentUUIDArrayField.mjs';
 export { default as MappingField } from './mappingField.mjs';
+export * as ActionFields from './action/_module.mjs';
