+++ module/data/fields/action/_module.mjs
@@ -2,6 +2,9 @@ export { default as CostField } from './costField.mjs';
 export { default as UsesField } from './usesField.mjs';
 export { default as RangeField } from './rangeField.mjs';
 export { default as TargetField } from './targetField.mjs';
-export { default as EffectField } from './effectField.mjs';
+export { default as EffectsField } from './effectsField.mjs';
 export { default as SaveField } from './saveField.mjs';
 export { default as BeastformField } from './beastformField.mjs';
+export { default as DamageField } from './damageField.mjs';
+export { default as HealingField } from './healingField.mjs';
+export { default as RollField } from './rollField.mjs';
