+++ module/data/fields/action/effectsField.mjs
@@ -1,6 +1,6 @@
 const fields = foundry.data.fields;
 
-export default class EffectField extends fields.ArrayField {
+export default class EffectsField extends fields.ArrayField {
     constructor(options={}, context={}) {
         const element = new fields.SchemaField({
             _id: new fields.DocumentIdField(),
