+++ module/data/fields/action/rangeField.mjs
@@ -9,4 +9,6 @@ export default class RangeField extends fields.StringField {
         };
         super(options, context);
     }
+
+    static prepareConfig(config) {}
 }
